+++ src/Streaming/Linear.hs
@@ -63,7 +63,7 @@ import Prelude.Linear (($), (.), (&))
 import Prelude (Ordering(..), Ord(..), Num(..), Int, Either(..), Double,
                Maybe(..), fromInteger)
 import qualified Prelude
-import qualified Control.Monad.Linear as Control
+import qualified Control.Functor.Linear as Control
 import qualified Data.Functor.Linear as Data
 import Data.Unrestricted.Linear
 import Control.Concurrent (threadDelay)
